refactor: 重构项目结构并迁移至React
- 将项目从原生JavaScript迁移至React框架
- 重新组织文件结构，将资源文件移至assets目录
- 添加Tailwind CSS和PostCSS配置
- 更新构建配置，添加TypeScript支持
- 删除旧有的HTML和JavaScript文件
- 优化代码结构，提升可维护性
