--- manifest.json
@@ -1,42 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "穷学宝典自动答题插件",
-  "version": "1.0",
-  "description": "自动答题并记录正确答案，支持防止视频暂停功能",
-  "permissions": [
-    "activeTab",
-    "storage",
-    "scripting",
-    "tabs"
-  ],
-  "host_permissions": [
-    "https://iedu.foxconn.com/*",
-    "http://iedu.foxconn.com/*",
-    "http://video.fulearn.com/*",
-    "https://video.fulearn.com/*"
-  ],
-  "background": {
-    "service_worker": "background.js"
-  },
-  "content_scripts": [
-    {
-      "matches": [
-        "https://iedu.foxconn.com/*", 
-        "http://iedu.foxconn.com/*",
-        "http://video.fulearn.com/*", 
-        "https://video.fulearn.com/*",
-        "*://iedu.foxconn.com/public/user/playCourse?*",
-        "*://iedu.foxconn.com/public/play/play*"
-      ],
-      "js": ["content.js"]
-    }
-  ],
-  "action": {
-    "default_popup": "popup.html",
-    "default_icon": {
-      "16": "icons/icon16.png",
-      "48": "icons/icon48.png",
-      "128": "icons/icon128.png"
-    }
-  }
-}
