+++ postcss.config.js
@@ -1,0 +1,8 @@
+/**
+ * @type {import('postcss').ProcessOptions}
+ */
+module.exports = {
+  plugins: {
+    tailwindcss: {}
+  }
+}
