+++ tailwind.config.js
@@ -1,0 +1,7 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: ["./src/**/*.{tsx,html}"],
+  darkMode: "media",
+  // 移除或修改 prefix 配置
+  prefix: "" // 或改为合适的前缀
+}
